modify config structure

+++ profile/cmd/server.go
@@ -92,10 +92,9 @@ func (p ProfileServer) ListAccounts(ctx context.Context, request *v1.ListAccount
 	}
 	return &v1.ListAccount{Account: findAccounts}, nil
 }
-
 func (p ProfileServer) DeleteAccount(ctx context.Context, request *v1.AccountRequest) (*empty.Empty, error) {
-	if request.AccountId == "" {
-		return nil, errors.New("account_id is required")
+	if request == nil || request.AccountId == "" {
+		return nil, status.Error(codes.InvalidArgument, "account_id is required")
 	}
 	err := p.account.DeleteAccount(ctx, account.ProtoToAccountRequest(request))
 	if err != nil {
@@ -106,7 +105,6 @@ func (p ProfileServer) DeleteAccount(ctx context.Context, request *v1.AccountReq
 	}
 	return nil, nil
 }
-
 func (p ProfileServer) CreateUser(ctx context.Context, request *v1.User) (*empty.Empty, error) {
 	_, err := p.user.CreateUser(user.ProtoToUser(request))
 	if err != nil {
